chore(wallet): fix animation layout
