+++ wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/transfer/TransferSheet.kt
@@ -1,24 +1,16 @@
 package eu.peernetwork.wallet.ui.transfer
 
 import android.content.res.Configuration
-import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.Crossfade
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.derivedStateOf
@@ -26,15 +18,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -49,10 +37,8 @@ import eu.peernetwork.wallet.ui.model.UiRecipient
 import eu.peernetwork.wallet.ui.model.UiTransfer
 import java.math.BigDecimal
 import java.util.UUID
-import com.airbnb.lottie.compose.LottieAnimation
-import com.airbnb.lottie.compose.LottieCompositionSpec
-import com.airbnb.lottie.compose.animateLottieCompositionAsState
-import com.airbnb.lottie.compose.rememberLottieComposition
+import eu.peernetwork.wallet.ui.compose.TickButton
+import eu.peernetwork.wallet.ui.compose.ToastLayout
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
@@ -150,26 +136,23 @@ fun TransferSheet(
     Crossfade(isSuccessful.value) { target ->
         if (target) {
             Box {
-            TransferSheetScaffold(
-                state = state,
-                error = error,
-                title = stringResource(R.string.sent_label),
-                recipient = recipient,
-                token = token,
-                action = stringResource(R.string.close_label),
-                onClick = onClick,
-                onSubmit = onSubmit,
-            ) {
-                LottieTickIcon(
-                    playKey = true,
-                    modifier = Modifier.size(36.dp),
-                    rawRes = R.raw.tick
-                )
+                TransferSheetScaffold(
+                    state = state,
+                    error = error,
+                    title = stringResource(R.string.sent_label),
+                    recipient = recipient,
+                    token = token,
+                    action = stringResource(R.string.close_label),
+                    onClick = onClick,
+                    onSubmit = onSubmit,
+                ) {
+                    TickButton(
+                        modifier = Modifier.size(36.dp),
+                        rawRes = R.raw.tick
+                    )
+                }
+                ToastLayout()
             }
-                TransferSuccessOverlay(
-                    playKey = true,
-                    rawRes = R.raw.tick
-                )}
         } else {
             TransferSheetScaffold(
                 state = state,
@@ -192,99 +175,6 @@ fun TransferSheet(
     }
 }
 
-@Composable
-private fun LottieTickIcon(
-    playKey: Boolean,
-    modifier: Modifier = Modifier,
-    rawRes: Int = R.raw.tick
-) {
-    var playing by remember(playKey) { mutableStateOf(false) }
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(rawRes))
-    val progress by animateLottieCompositionAsState(
-        composition = composition,
-        isPlaying = playing,
-        iterations = 1,
-        speed = 1f
-    )
-    LaunchedEffect(playKey) {
-        if (playKey) {
-            playing = false
-            playing = true
-        }
-    }
-    LottieAnimation(
-        composition = composition,
-        progress = { progress },
-        modifier = modifier
-    )
-}
-
-@Composable
-private fun BoxScope.TransferSuccessOverlay(
-    playKey: Boolean,
-    modifier: Modifier = Modifier,
-    size: Dp = 320.dp,
-    rawRes: Int = R.raw.peerrocket,
-    fillToWidthEdges: Boolean = true,
-    overscanScale: Float = 1.12f,
-    anchorBottom: Boolean = true
-) {
-    var play by remember { mutableStateOf(false) }
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(rawRes))
-    val progress by animateLottieCompositionAsState(
-        composition = composition,
-        isPlaying = play,
-        iterations = 1,
-        speed = 1f
-    )
-
-    LaunchedEffect(playKey) {
-        if (playKey) {
-            play = false
-            play = true
-        }
-    }
-    LaunchedEffect(progress) {
-        if (progress >= 1f) play = false
-    }
-
-    val aspect: Float = remember(composition) {
-        val b = composition?.bounds
-        val w = b?.width()?.toFloat() ?: 1f
-        val h = b?.height()?.toFloat() ?: 1f
-        if (h > 0f) w / h else 1f
-    }
-
-    AnimatedVisibility(
-        visible = play && progress < 1f,
-        enter = fadeIn(animationSpec = tween(160)),
-        exit = fadeOut(animationSpec = tween(220)),
-        modifier = modifier.matchParentSize()
-    ) {
-        Box(
-            modifier = Modifier.matchParentSize(),
-            contentAlignment = if (anchorBottom) Alignment.BottomCenter else Alignment.Center
-        ) {
-            val lottieMod = if (fillToWidthEdges) {
-                Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(aspect, matchHeightConstraintsFirst = false)
-                    .graphicsLayer {
-                        scaleX = overscanScale
-                        scaleY = overscanScale
-                    }
-            } else {
-                Modifier.size(size)
-            }
-
-            LottieAnimation(
-                composition = composition,
-                progress = { progress },
-                modifier = lottieMod
-            )
-        }
-    }
-}
 @Composable
 @Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 fun PreviewTransferSheet() {
